Add provider lookup and related utils to shared
